move function random_int to util
